Added method for getting subject with Lessons entities and changed Id from int to long in Subject entity

+++ EStudy/EStudy.Domain/Models/Lesson.cs
@@ -23,7 +23,7 @@ namespace EStudy.Domain.Models
         public int? TeacherReplacementId { get; set; }
         [Required]
         public TypeLesson TypeLesson { get; set; }
-        public int SubjectId { get; set; }
+        public long SubjectId { get; set; }
         public Subject Subject { get; set; }
         public List<Task> Tasks { get; set; }
         public List<Mark> Marks { get; set; }
